feat: functionality

+++ app/src/main/java/com/example/mefit/model/Food.kt
@@ -2,9 +2,9 @@ package com.example.mefit.model
 
 
 data class Food(
-    var id: Int,
+    var id: Long,
     var name: String,
-    var calories: Int,
+    var calories: Long,
 )
